:sparkles: feat: TMap 에서 자가용 교통 정보 가져오기
Issue #85

+++ data/src/main/kotlin/com/woory/data/model/PathModel.kt
@@ -2,6 +2,6 @@ package com.woory.data.model
 
 data class PathModel(
     val routeType: RouteType,
-    val distance: Double,
+    val distance: Int,
     val time: Int,
 )
